#0: Adding comments to sharded addrgen and cleaning up what function is in the api_common
